old-tests/submission/Microsoft/selection/select.htm

<!DOCTYPE HTML>
<html>
    <head>
        <title id="desc">HTML5 Selection: Call select() on a text field</title>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="common.js"></script>
        <script type="text/javascript">
            function RunTest()
            {
                var input1 = document.getElementById("input1");

                input1.select();

                checkDefaultSelectionAttributes();

                var selectionText = input1.value.substring(input1.selectionStart, input1.selectionEnd);
                assert_equals(selectionText, input1.value);
            }
        </script>
    </head>
    <body onload="test(RunTest);">
        <input style="WIDTH: 500px" id="input1" value="Some text in an input control" type="text" />
        <p>Select the text in the input element by calling select()</p>
    </body>
</html>
